Refactor Loading component
